Added some Logos and user related data - - - > still working on the project

+++ library/books/views.py
@@ -3,7 +3,6 @@ from .forms import BookRegistration
 from .models import Book
 from django.contrib import messages
 
-# Create your views here.
 
 Book_title = []
 
@@ -16,31 +15,36 @@ def home(request):
 
 
 def add_book(request):
-    if request.method == 'POST':
-        form_ = BookRegistration(request.POST)
-        if form_.is_valid():
-            tl = form_.cleaned_data['title']
-            au = form_.cleaned_data['author']
-            pr = form_.cleaned_data['price']
-            pd = form_.cleaned_data['publication_date']
-            isbn = form_.cleaned_data['ISBN_no']
-            b_data = Book(title=tl, author=au, price=pr,
-                          publication_date=pd, ISBN_no=isbn)
-            b_data.save()
-            Book_title.append(tl)
-            ##########
+    if request.user.is_superuser:
+        if request.method == 'POST':
+            form_ = BookRegistration(request.POST)
+            if form_.is_valid():
+                tl = form_.cleaned_data['title']
+                au = form_.cleaned_data['author']
+                pr = form_.cleaned_data['price']
+                pd = form_.cleaned_data['publication_date']
+                isbn = form_.cleaned_data['ISBN_no']
+                b_data = Book(title=tl, author=au, price=pr,
+                              publication_date=pd, ISBN_no=isbn)
+                b_data.save()
+                Book_title.append(tl)
+                ##########
 
-            if len(Book_title) == 0:
-                return None
-            else:
-                print(Book_title)
-            messages.warning(
-                request, f'{tl} is successfully added!')
-            return redirect('show_book')
+                if len(Book_title) == 0:
+                    return None
+                else:
+                    print(Book_title)
+                messages.warning(
+                    request, f'{tl} is successfully added!')
+                return redirect('show_book')
 
+        else:
+            form_ = BookRegistration()
+        return render(request, 'books/add.html', {'form': form_})
     else:
-        form_ = BookRegistration()
-    return render(request, 'books/add.html', {'form': form_})
+        messages.warning(
+            request, "Only Admin Have Permission To Add A Book!")
+        return redirect('home')
 
 
 def show_book(request):
@@ -49,31 +53,47 @@ def show_book(request):
 
 
 def update_book(request):
-    book_list = Book.objects.all()
-    return render(request, 'books/update.html', {'book_data': book_list})
+    if request.user.is_superuser:
+        book_list = Book.objects.all()
+        return render(request, 'books/update.html', {'book_data': book_list})
+    else:
+        messages.warning(
+            request, "Only Admin Have Permission To Update A Book!")
+        return redirect('home')
 
 
 def delete_book(request, id):
-    if request.method == "POST":
-        obj = Book.objects.get(pk=id)
-        print(obj.title)
-        obj.delete()
-        messages.info(
-            request, f'Book {obj.title} is successfully deleted!')
-        return redirect('update_book')
+    if request.user.is_superuser:
+        if request.method == "POST":
+            obj = Book.objects.get(pk=id)
+            print(obj.title)
+            obj.delete()
+            messages.info(
+                request, f'Book {obj.title} is successfully deleted!')
+            return redirect('update_book')
+    else:
+        messages.warning(
+            request, "Only Admin Have Permission To Delete A Book!")
+        return redirect('home')
 
 
 def edit_book(request, id):
-    if request.method == 'POST':
-        obj = Book.objects.get(pk=id)
-        form_ = BookRegistration(request.POST, instance=obj)
-        if form_.is_valid():
-            tl = form_.cleaned_data['title']
-            form_.save()
-            messages.warning(
-                request, f'Data of {tl} is successfully updated!')
-            return redirect('show_book')
+
+    if request.user.is_superuser:
+        if request.method == 'POST':
+            obj = Book.objects.get(pk=id)
+            form_ = BookRegistration(request.POST, instance=obj)
+            if form_.is_valid():
+                tl = form_.cleaned_data['title']
+                form_.save()
+                messages.warning(
+                    request, f'Data of {tl} is successfully updated!')
+                return redirect('show_book')
+        else:
+            obj = Book.objects.get(pk=id)
+            form_ = BookRegistration(instance=obj)
+        return render(request, 'books/edit_book.html', {'form': form_})
     else:
-        obj = Book.objects.get(pk=id)
-        form_ = BookRegistration(instance=obj)
-    return render(request, 'books/edit_book.html', {'form': form_})
+        messages.warning(
+            request, "Only Admin Have Permission To Update A Book!")
+        return redirect('home')
